Rename and move things around. Pass API URL from outside of the component
