+++ docs/pages/x/api/data-grid/grid-actions-col-def.json
@@ -36,6 +36,7 @@
     "disableReorder": { "type": { "description": "boolean" }, "default": "false" },
     "display": { "type": { "description": "'text' | 'flex'" } },
     "editable": { "type": { "description": "boolean" }, "default": "false" },
+    "examples": { "type": { "description": "V[]" } },
     "filterable": { "type": { "description": "boolean" }, "default": "true" },
     "filterOperators": {
       "type": { "description": "readonly GridFilterOperator&lt;R, V, F&gt;[]" }
