+++ docs/pages/x/api/data-grid/grid-single-select-col-def.json
@@ -34,6 +34,7 @@
     "disableReorder": { "type": { "description": "boolean" }, "default": "false" },
     "display": { "type": { "description": "'text' | 'flex'" } },
     "editable": { "type": { "description": "boolean" }, "default": "false" },
+    "examples": { "type": { "description": "V[]" } },
     "filterable": { "type": { "description": "boolean" }, "default": "true" },
     "filterOperators": {
       "type": { "description": "readonly GridFilterOperator&lt;R, V, F&gt;[]" }
