+++ docs/translations/api-docs/data-grid/data-grid-premium/data-grid-premium.json
@@ -1199,11 +1199,27 @@
       "description": "Styles applied to {{nodeName}}.",
       "nodeName": "the toolbar filter list element"
     },
-    "toolbarRemoteControl": { "description": "" },
-    "toolbarRemoteControl--recording": { "description": "" },
-    "toolbarRemoteControlRecordButton": { "description": "" },
-    "toolbarRemoteControlRecordingIndicator": { "description": "" },
-    "toolbarRemoteControlSendButton": { "description": "" },
+    "toolbarPromptControl": {
+      "description": "Styles applied to {{nodeName}}.",
+      "nodeName": "the toolbar prompt control element"
+    },
+    "toolbarPromptControl--recording": {
+      "description": "Styles applied to {{nodeName}} when {{conditions}}.",
+      "nodeName": "the toolbar prompt control element",
+      "conditions": "recording"
+    },
+    "toolbarPromptControlRecordButton": {
+      "description": "Styles applied to {{nodeName}}.",
+      "nodeName": "the toolbar prompt control record button element"
+    },
+    "toolbarPromptControlRecordingIndicator": {
+      "description": "Styles applied to {{nodeName}}.",
+      "nodeName": "the toolbar prompt control recording indicator element"
+    },
+    "toolbarPromptControlSendButton": {
+      "description": "Styles applied to {{nodeName}}.",
+      "nodeName": "the toolbar prompt control send button element"
+    },
     "treeDataGroupingCell": {
       "description": "Styles applied to {{nodeName}}.",
       "nodeName": "the root of the grouping column of the tree data"
