+++ docs/translations/api-docs/data-grid/grid-single-select-col-def.json
@@ -28,6 +28,9 @@
       "description": "Display mode for the cell:<br /> - &#39;text&#39;: For text-based cells (default)<br /> - &#39;flex&#39;: For cells with HTMLElement children"
     },
     "editable": { "description": "If <code>true</code>, the cells of the column are editable." },
+    "examples": {
+      "description": "Example values that can be used by the grid to get more context about the column."
+    },
     "filterable": { "description": "If <code>true</code>, the column is filterable." },
     "filterOperators": { "description": "Allows setting the filter operators for this column." },
     "flex": { "description": "If set, it indicates that a column has fluid width. Range [0, ∞)." },
