+++ packages/x-data-grid-premium/src/components/index.ts
@@ -1,5 +1,5 @@
 export * from './GridExcelExportMenuItem';
 export * from '../material/icons';
 export * from './GridColumnMenuAggregationItem';
-export * from './GridToolbarRemoteControl';
+export * from './promptControl';
 export { GridColumnMenuGroupingItem } from './GridPremiumColumnMenu';
