+++ packages/x-data-grid-premium/src/components/promptControl/index.ts
@@ -1,0 +1,1 @@
+export * from './GridToolbarPromptControl';
