+++ packages/x-data-grid/src/constants/gridClasses.ts
@@ -592,11 +592,26 @@ export interface GridClasses {
    * Styles applied to the toolbar filter list element.
    */
   toolbarFilterList: string;
-  toolbarRemoteControl: string;
-  'toolbarRemoteControl--recording': string;
-  toolbarRemoteControlRecordingIndicator: string;
-  toolbarRemoteControlRecordButton: string;
-  toolbarRemoteControlSendButton: string;
+  /**
+   * Styles applied to the toolbar prompt control element.
+   */
+  toolbarPromptControl: string;
+  /**
+   * Styles applied to the toolbar prompt control element when recording.
+   */
+  'toolbarPromptControl--recording': string;
+  /**
+   * Styles applied to the toolbar prompt control recording indicator element.
+   */
+  toolbarPromptControlRecordingIndicator: string;
+  /**
+   * Styles applied to the toolbar prompt control record button element.
+   */
+  toolbarPromptControlRecordButton: string;
+  /**
+   * Styles applied to the toolbar prompt control send button element.
+   */
+  toolbarPromptControlSendButton: string;
 
   /**
    * Styles applied the grid if `showColumnVerticalBorder={true}`.
@@ -809,11 +824,11 @@ export const gridClasses = generateUtilityClasses<GridClassKey>('MuiDataGrid', [
   'sortIcon',
   'toolbarContainer',
   'toolbarFilterList',
-  'toolbarRemoteControl',
-  'toolbarRemoteControl--recording',
-  'toolbarRemoteControlRecordingIndicator',
-  'toolbarRemoteControlRecordButton',
-  'toolbarRemoteControlSendButton',
+  'toolbarPromptControl',
+  'toolbarPromptControl--recording',
+  'toolbarPromptControlRecordingIndicator',
+  'toolbarPromptControlRecordButton',
+  'toolbarPromptControlSendButton',
   'virtualScroller',
   'virtualScroller--hasScrollX',
   'virtualScrollerContent',
