+++ scripts/x-data-grid-premium.exports.json
@@ -604,11 +604,11 @@
   { "name": "GridToolbarExportProps", "kind": "Interface" },
   { "name": "GridToolbarFilterButton", "kind": "Variable" },
   { "name": "GridToolbarFilterButtonProps", "kind": "Interface" },
+  { "name": "GridToolbarPromptControl", "kind": "Function" },
+  { "name": "GridToolbarPromptControlProps", "kind": "TypeAlias" },
   { "name": "GridToolbarProps", "kind": "Interface" },
   { "name": "GridToolbarQuickFilter", "kind": "Function" },
   { "name": "GridToolbarQuickFilterProps", "kind": "TypeAlias" },
-  { "name": "GridToolbarRemoteControl", "kind": "Function" },
-  { "name": "GridToolbarRemoteControlProps", "kind": "TypeAlias" },
   { "name": "gridTopLevelRowCountSelector", "kind": "Variable" },
   { "name": "GridTranslationKeys", "kind": "TypeAlias" },
   { "name": "GridTreeBasicNode", "kind": "Interface" },
